TestServer: Put common code for starting the server in TestServerData
