+++ gotham/src/tls/test.rs
@@ -6,7 +6,7 @@ use std::future::Future;
 use std::io::{self, BufReader};
 use std::net::{self, SocketAddr};
 use std::pin::Pin;
-use std::sync::{Arc, RwLock};
+use std::sync::Arc;
 use std::task::{Context, Poll};
 use std::time::Duration;
 
@@ -19,9 +19,7 @@ use log::info;
 use pin_project::pin_project;
 use rustls::Session;
 use tokio::io::{AsyncRead, AsyncWrite, ReadBuf};
-use tokio::net::TcpListener;
 use tokio::net::TcpStream;
-use tokio::runtime::Runtime;
 use tokio::time::{sleep, Sleep};
 use tokio_rustls::client::TlsStream;
 use tokio_rustls::{
@@ -104,20 +102,7 @@ impl TestServer {
         new_handler: NH,
         timeout: u64,
     ) -> anyhow::Result<TestServer> {
-        let runtime = Runtime::new()?;
-        // TODO: Fix this into an async flow
-        let listener = runtime.block_on(TcpListener::bind("127.0.0.1:0".parse::<SocketAddr>()?))?;
-        let addr = listener.local_addr()?;
-
-        let wrap = create_wrap()?;
-        let service_stream = super::bind_server(listener, new_handler, wrap);
-        runtime.spawn(service_stream); // Ignore the result
-
-        let data = TestServerData {
-            addr,
-            timeout,
-            runtime: RwLock::new(runtime),
-        };
+        let data = TestServerData::new(new_handler, timeout, create_wrap()?)?;
 
         Ok(TestServer {
             data: Arc::new(data),
